# Conflicts:
#	CMakeLists.txt
#	Common/core/def_version.h
#	Editor/AGS.Editor/AGSEditor.csproj
#	Editor/AGS.Editor/Components/AudioComponent.cs
#	Editor/AGS.Editor/Components/FontsComponent.cs
#	Editor/AGS.Editor/Components/RoomsComponent.cs
#	Editor/AGS.Editor/Panes/Room/RoomEditFilters/ObjectsEditorFilter.cs
#	Editor/AGS.Editor/Panes/Room/RoomSettingsEditor.cs
#	Editor/AGS.Editor/Panes/SpriteImportWindow.cs
#	Editor/AGS.Editor/Panes/SpriteSelector.cs
#	Editor/AGS.Editor/Properties/Resources.resx
#	Editor/AGS.Editor/Resources/agsdefns.sh
#	Editor/AGS.Editor/Tasks.cs
#	Editor/AGS.Editor/Utils/SpriteTools.cs
#	Editor/AGS.Editor/Utils/Utilities.cs
#	Editor/AGS.Editor/app.manifest
#	Editor/AGS.Types/AGS.Types.csproj
#	Editor/AGS.Types/AudioClip.cs
#	Editor/AGS.Types/Dialog.cs
#	Editor/AGS.Types/Properties/AssemblyInfo.cs
#	Editor/AGS.Types/RoomWalkBehind.cs
#	Editor/AGS.Types/RoomWalkableArea.cs
#	Engine/ac/character.cpp
#	Engine/ac/character.h
#	Engine/ac/display.cpp
#	Engine/ac/dynobj/dynobj_manager.h
#	Engine/ac/global_api.cpp
#	Engine/ac/global_game.cpp
#	Engine/ac/global_game.h
#	Engine/ac/global_object.cpp
#	Engine/ac/global_object.h
#	Engine/ac/gui.cpp
#	Engine/ac/room.cpp
#	Engine/ac/route_finder.cpp
#	Engine/device/mousew32.cpp
#	Engine/main/game_run.cpp
#	Engine/platform/windows/acplwin.cpp
#	Engine/platform/windows/gfx/ali3dd3d.cpp
#	Engine/plugin/agsplugin.cpp
#	Engine/script/cc_instance.h
#	Engine/script/scriptexecutor.cpp
#	Plugins/agspalrender/agspalrender/ags_palrender.cpp
#	version.json

+++ Engine/ac/draw.h
@@ -167,7 +167,7 @@ void draw_gui_sprite_flipped(Common::Bitmap *ds, int xpos, int ypos,
 // Render game on screen
 void render_to_screen();
 void GfxDriverOnInitCallback(void *data);
-bool GfxDriverSpriteEvtCallback(int evt, int data);
+bool GfxDriverSpriteEvtCallback(int evt, intptr_t data);
 
 // Create the actsps[objid] image with the object drawn correctly.
 // Returns true if nothing at all has changed and actsps is still
