+++ Engine/plugin/agsplugin.h
@@ -488,7 +488,7 @@ public:
   // check whether a script function can be run now
   AGSIFUNC(int)    CanRunScriptFunctionNow();
   // call a user-defined script function
-  AGSIFUNC(int)    CallGameScriptFunction(const char *name, int32 globalScript, int32 numArgs, long arg1 = 0, long arg2 = 0, long arg3 = 0);
+  AGSIFUNC(int)    CallGameScriptFunction(const char *name, int32 globalScript, int32 numArgs, intptr_t arg1 = 0, intptr_t arg2 = 0, intptr_t arg3 = 0);
 
   // *** BELOW ARE INTERFACE VERSION 15 AND ABOVE ONLY
   // force any sprites on-screen using the slot to be updated
@@ -496,7 +496,7 @@ public:
   // change whether the specified sprite is a 32-bit alpha blended image
   AGSIFUNC(void)   SetSpriteAlphaBlended(int32 slot, int32 isAlphaBlended);
   // run the specified script function whenever script engine is available
-  AGSIFUNC(void)   QueueGameScriptFunction(const char *name, int32 globalScript, int32 numArgs, long arg1 = 0, long arg2 = 0);
+  AGSIFUNC(void)   QueueGameScriptFunction(const char *name, int32 globalScript, int32 numArgs, intptr_t arg1 = 0, intptr_t arg2 = 0);
   // register a new dynamic managed script object
   AGSIFUNC(int)    RegisterManagedObject(void *object, IAGSScriptManagedObject *callback);
   // add an object reader for the specified object type
@@ -652,7 +652,7 @@ DLLEXPORT int    AGS_EditorSaveGame (char *, int);
 DLLEXPORT void   AGS_EditorLoadGame (char *, int);
 DLLEXPORT void   AGS_EngineStartup (IAGSEngine *);
 DLLEXPORT void   AGS_EngineShutdown (void);
-DLLEXPORT int    AGS_EngineOnEvent (int, int);
+DLLEXPORT intptr_t AGS_EngineOnEvent (int, intptr_t);
 DLLEXPORT int    AGS_EngineDebugHook(const char *, int, int);
 DLLEXPORT void   AGS_EngineInitGfx(const char* driverID, void *data); 
 // Export this to let engine verify that this is a compatible AGS Plugin;
