+++ Plugins/ags_parallax/ags_parallax/ags_parallax.cpp
@@ -214,7 +214,7 @@ void AGS_EngineShutdown()
 {
 }
 
-int AGS_EngineOnEvent(int event, int data)
+intptr_t AGS_EngineOnEvent(int event, intptr_t data)
 {
   if (event == AGSE_PREGUIDRAW)
   {
